refactor: rename exceptions directory to errors
